cleanMetadata
clean sample metadata, prepare for use of namespaces so rosetta can use
the namespaces we assigned in all crosswalks, recreate some sample data
csv that people can use to compare collections,

Also updates to the notebook that processes the metadata. there is now
a version that interfaces with the evaluator service, allowing people
to create a directory structure, place their metadata in it and prepare
it for the service.